Check if the pawn is too big for the bed.
This is not really a relevant situation in vanilla, as the ways to
modify a pawn's body size are limited, and non-animal beds do not
limit pawns by body size. However, it can come into play with mods
that allow increasing body size greatly, and other mods that limit the
body size a bed can hold.